refactor(engines): split pulid_flux1_dev.py into _pulid/ sub-package (#55)
Extract PuLID nn.Module definitions and FLUX.1-dev transformer patching
out of pulid_flux1_dev.py (569 LOC) into a private _pulid/ sub-package
so the engine file clears the 300 LOC quality gate. Behavior unchanged.

- _pulid/modules.py: IDFormer, PerceiverAttention, PerceiverAttentionCA,
  PuLIDFlux1 container with from_safetensors classmethod
- _pulid/patching.py: patch_flux1() monkey-patcher + unpatch callable
- engines/pulid_flux1_dev.py: now 224 LOC, consumes the sub-package
- tools/file_exemptions.txt: drop the #55 exemption

Golden-image regression test (T0/T7 in the plan) deferred — requires GPU
and real PuLID weights. Manual smoke-test with a known seed before merge.
